Rename staging module k8s.io/webhook-auth to k8s.io/webhookauth
Drop the hyphen for consistency with sibling staging module names, which use a single lowercase token. Mechanical rename only: directory move, module path, intra-module import paths, go.work, the root go.mod replace directive, and the publishing rules entry. No behavior change.
