+++ go.work
@@ -39,5 +39,5 @@ use (
 	./staging/src/k8s.io/sample-cli-plugin
 	./staging/src/k8s.io/sample-controller
 	./staging/src/k8s.io/streaming
-	./staging/src/k8s.io/webhook-auth
+	./staging/src/k8s.io/webhookauth
 )
