+++ staging/src/k8s.io/webhookauth/verify/admissionhttp/admissionhttp_test.go
@@ -43,9 +43,9 @@ import (
 	admissionv1 "k8s.io/api/admission/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/webhook-auth/verify"
-	"k8s.io/webhook-auth/verify/admissionhttp"
-	"k8s.io/webhook-auth/verify/oidc"
+	"k8s.io/webhookauth/verify"
+	"k8s.io/webhookauth/verify/admissionhttp"
+	"k8s.io/webhookauth/verify/oidc"
 )
 
 const (
