+++ staging/src/k8s.io/webhookauth/verify/admissionhttp/audience_enforcement_test.go
@@ -28,9 +28,9 @@ import (
 
 	admissionv1 "k8s.io/api/admission/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/webhook-auth/verify"
-	"k8s.io/webhook-auth/verify/admissionhttp"
-	"k8s.io/webhook-auth/verify/oidc"
+	"k8s.io/webhookauth/verify"
+	"k8s.io/webhookauth/verify/admissionhttp"
+	"k8s.io/webhookauth/verify/oidc"
 )
 
 // TestAudienceFromServiceURL_ParseError covers the parse-failure branch: a
