+++ staging/src/k8s.io/webhookauth/verify/admissionhttp/failclosed_test.go
@@ -27,8 +27,8 @@ import (
 
 	admissionv1 "k8s.io/api/admission/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/webhook-auth/verify"
-	"k8s.io/webhook-auth/verify/admissionhttp"
+	"k8s.io/webhookauth/verify"
+	"k8s.io/webhookauth/verify/admissionhttp"
 )
 
 // TestBearerToken_Table exercises the header parser directly across the scheme
