+++ staging/src/k8s.io/webhookauth/verify/oidc/oidc_test.go
@@ -29,8 +29,8 @@ import (
 
 	jose "gopkg.in/go-jose/go-jose.v2"
 
-	"k8s.io/webhook-auth/verify"
-	"k8s.io/webhook-auth/verify/oidc"
+	"k8s.io/webhookauth/verify"
+	"k8s.io/webhookauth/verify/oidc"
 )
 
 const (
