+++ staging/src/k8s.io/webhookauth/verify/oidc/rotation_test.go
@@ -29,7 +29,7 @@ import (
 
 	jose "gopkg.in/go-jose/go-jose.v2"
 
-	"k8s.io/webhook-auth/verify/oidc"
+	"k8s.io/webhookauth/verify/oidc"
 )
 
 // rotatingOIDCServer is a TLS OIDC discovery + JWKS endpoint whose signing key
